* Source/NSMessagePort.m: * Source/NSFileManager.m: * Source/NSCharacterSet.m: Fix some method type mismatches. Generally when a private subclass overrides a public method taking a NSUInteger with a method taking an "unsigned" parameter.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@31351 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSCharacterSet.m
@@ -77,7 +77,7 @@
 /* Private class from NSIndexSet.m
  */
 @interface	_GSStaticIndexSet : NSIndexSet
-- (id) _initWithBytes: (const void*)bytes length: (unsigned)length;
+- (id) _initWithBytes: (const void*)bytes length: (NSUInteger)length;
 @end
 
 @interface NSBitmapCharSet : NSCharacterSet
@@ -1212,7 +1212,7 @@ static Class concreteMutableClass = nil;
  */
 
 @interface      NSIndexSet (NSCharacterSet)
-- (unsigned int) _gapGreaterThanIndex: (unsigned int)anIndex;
+- (NSUInteger) _gapGreaterThanIndex: (NSUInteger)anIndex;
 @end
